Move humility_cmd::i2c to a crate; consolidate humility_cmd::test
