+++ cmd/i2c/Cargo.toml
@@ -15,3 +15,4 @@ log.workspace = true
 humility.workspace = true
 humility-cmd.workspace = true
 humility-hiffy.workspace = true
+humility-i2c.workspace = true
