+++ cmd/pmbus/Cargo.toml
@@ -17,3 +17,4 @@ parse_int.workspace = true
 humility.workspace = true
 humility-cmd.workspace = true
 humility-hiffy.workspace = true
+humility-i2c.workspace = true
