+++ cmd/pmbus/src/lib.rs
@@ -6,9 +6,9 @@ use colored::Colorize;
 use humility::cli::Subcommand;
 use humility::core::Core;
 use humility::hubris::*;
-use humility_cmd::i2c::I2cArgs;
 use humility_cmd::{Archive, Attach, Command, CommandKind, Validate};
 use humility_hiffy::*;
+use humility_i2c::I2cArgs;
 
 use anyhow::{bail, Result};
 use clap::{CommandFactory, Parser};
