+++ cmd/rencm/Cargo.toml
@@ -19,3 +19,4 @@ csv.workspace = true
 humility.workspace = true
 humility-cmd.workspace = true
 humility-hiffy.workspace = true
+humility-i2c.workspace = true
