+++ cmd/rendmp/src/lib.rs
@@ -126,9 +126,9 @@
 
 use humility::cli::Subcommand;
 use humility::hubris::*;
-use humility_cmd::i2c::I2cArgs;
 use humility_cmd::{Archive, Attach, Command, CommandKind, Validate};
 use humility_hiffy::*;
+use humility_i2c::I2cArgs;
 use humility_idol::{HubrisIdol, IdolArgument};
 
 use anyhow::{anyhow, bail, Context, Result};
