+++ cmd/validate/Cargo.toml
@@ -18,3 +18,4 @@ humility.workspace = true
 humility-cmd.workspace = true
 humility-hiffy.workspace = true
 humility-idol.workspace = true
+humility-i2c.workspace = true
