+++ cmd/validate/src/lib.rs
@@ -48,9 +48,9 @@ use colored::Colorize;
 use hif::*;
 use humility::cli::Subcommand;
 use humility::hubris::*;
-use humility_cmd::i2c::I2cArgs;
 use humility_cmd::{Archive, Attach, Command, CommandKind, Validate};
 use humility_hiffy::HiffyContext;
+use humility_i2c::I2cArgs;
 use humility_idol::{self as idol, HubrisIdol};
 
 #[derive(Parser, Debug)]
