+++ humility-cmd/src/lib.rs
@@ -2,9 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-pub mod i2c;
-pub mod test;
-
 use anyhow::{bail, Result};
 use clap::Command as ClapCommand;
 use humility::cli::Cli;
